requirements.txt
asgiref==3.8.1
Django==4.1.13
djongo==1.3.6
dnspython==2.7.0
mongoengine==0.29.1
pymongo==3.12.3
pytz==2024.2
sqlparse==0.2.4
tzdata==2024.2
djangorestframework==3.14.0